Add check for CSRF

+++ application.py
@@ -73,25 +73,28 @@ def home():
 	items = session.query(Item).order_by(desc(Item.id)).limit(10)
 	return render_template('categories.html', categories=categories, items=items)
 
-@app.route('/categories/new', methods=['GET', 'POST'])
-def newCategory():
-	if 'email' not in login_session:
-		flash('You need to be logged in to perform this action.')
-		return redirect(url_for('login'))
-	if request.method == 'POST':
-		newCategory = Category(name = request.form['name'])
-		session.add(newCategory)
-		session.commit()
-		return redirect(url_for('showCategory', category_id=newCategory.id))
-	else:
-		return render_template('newCategory.html', categories=categories)
-
 @app.route('/categories/<int:category_id>/')
 @app.route('/categories/<int:category_id>/items/')
 def showCategory(category_id):
 	category = session.query(Category).filter_by(id = category_id).one()
 	items = session.query(Item).filter_by(category_id = category.id)
 	return render_template('showCategory.html', categories=categories, category=category, items=items)
+
+@app.route('/categories/new', methods=['GET', 'POST'])
+def newCategory():
+	if 'email' not in login_session:
+		flash('You need to be logged in to perform this action.')
+		return redirect(url_for('login'))
+	if request.method == 'POST':
+		if login_session['state'] == request.form['csrf']:
+			newCategory = Category(name = request.form['name'])
+			session.add(newCategory)
+			session.commit()
+			return redirect(url_for('showCategory', category_id=newCategory.id))
+		else:
+			abort(403)
+	else:
+		return render_template('newCategory.html', categories=categories)
 
 @app.route('/categories/<int:category_id>/edit/', methods=['GET', 'POST'])
 def editCategory(category_id):
@@ -100,10 +103,13 @@ def editCategory(category_id):
 		return redirect(url_for('login'))
 	category = session.query(Category).filter_by(id = category_id).one()
 	if request.method == 'POST':
-		category.name = request.form['name']
-		session.add(category)
-		session.commit()
-		return redirect(url_for('showCategory', category_id=category.id))
+		if login_session['state'] == request.form['csrf']:
+			category.name = request.form['name']
+			session.add(category)
+			session.commit()
+			return redirect(url_for('showCategory', category_id=category.id))
+		else:
+			abort(403)
 	else:
 		return render_template('editCategory.html', categories=categories, category=category)
 
@@ -114,11 +120,20 @@ def deleteCategory(category_id):
 		return redirect(url_for('login'))
 	category = session.query(Category).filter_by(id = category_id).one()
 	if request.method == 'POST':
-		session.delete(category)
-		session.commit()
-		return redirect(url_for('home'))
+		if login_session['state'] == request.form['csrf']:
+			session.delete(category)
+			session.commit()
+			return redirect(url_for('home'))
+		else:
+			abort(403)
 	else:
 		return render_template('deleteCategory.html', categories=categories, category=category)
+
+@app.route('/categories/<int:category_id>/items/<int:item_id>/')
+def showItem(category_id, item_id):
+	category = session.query(Category).filter_by(id = category_id).one()
+	item = session.query(Item).filter_by(id = item_id).one()
+	return render_template('showItem.html', categories=categories, category=category, item=item)
 
 @app.route('/categories/<int:category_id>/items/new/', methods=['GET', 'POST'])
 def newItem(category_id):
@@ -127,20 +142,17 @@ def newItem(category_id):
 		return redirect(url_for('login'))
 	category = session.query(Category).filter_by(id = category_id).one()
 	if request.method == 'POST':
-		newItem = Item(name = request.form['name'],
-					   category_id = request.form['category'],
-					   description = request.form['description'])
-		session.add(newItem)
-		session.commit()
-		return redirect(url_for('showItem', category_id=category.id, item_id=newItem.id))
+		if login_session['state'] == request.form['csrf']:
+			newItem = Item(name = request.form['name'],
+						   category_id = request.form['category'],
+						   description = request.form['description'])
+			session.add(newItem)
+			session.commit()
+			return redirect(url_for('showItem', category_id=category.id, item_id=newItem.id))
+		else:
+			abort(403)
 	else:
 		return render_template('newItem.html', categories=categories, category=category)
-
-@app.route('/categories/<int:category_id>/items/<int:item_id>/')
-def showItem(category_id, item_id):
-	category = session.query(Category).filter_by(id = category_id).one()
-	item = session.query(Item).filter_by(id = item_id).one()
-	return render_template('showItem.html', categories=categories, category=category, item=item)
 
 @app.route('/categories/<int:category_id>/items/<int:item_id>/edit/', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
@@ -150,12 +162,15 @@ def editItem(category_id, item_id):
 	category = session.query(Category).filter_by(id = category_id).one()
 	item = session.query(Item).filter_by(id = item_id).one()
 	if request.method == 'POST':
-		item.name = request.form['name']
-		item.category_id = request.form['category']
-		item.description = request.form['description']
-		session.add(item)
-		session.commit()
-		return redirect(url_for('showItem', category_id=category.id, item_id=item.id))
+		if login_session['state'] == request.form['csrf']:
+			item.name = request.form['name']
+			item.category_id = request.form['category']
+			item.description = request.form['description']
+			session.add(item)
+			session.commit()
+			return redirect(url_for('showItem', category_id=category.id, item_id=item.id))
+		else:
+			abort(403)
 	else:
 		return render_template('editItem.html', categories=categories, category=category, item=item)
 
@@ -167,10 +182,13 @@ def deleteItem(category_id, item_id):
 	category = session.query(Category).filter_by(id = category_id).one()
 	item = session.query(Item).filter_by(id = item_id).one()
 	if request.method == 'POST':
-		session.delete(item)
-		session.commit()
-		items = session.query(Item).filter_by(category_id = category.id)
-		return redirect(url_for('showCategory', category_id=category.id))
+		if login_session['state'] == request.form['csrf']:
+			session.delete(item)
+			session.commit()
+			items = session.query(Item).filter_by(category_id = category.id)
+			return redirect(url_for('showCategory', category_id=category.id))
+		else:
+			abort(403)
 	else:
 		return render_template('deleteItem.html', categories=categories, category=category, item=item)
